backend/pyproject.toml
[project]
name = "whatthefridge"
version = "0.1.0"
description = "Cooking agent application with LangGraph"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "langgraph==0.0.20",
    "langchain==0.1.0",
    "langchain-google-genai==0.0.6",
    "google-generativeai==0.3.2",
    "youtube-search-python==1.6.6",
    "youtube-transcript-api==0.6.1",
    "opencv-python==4.8.1.78",
    "pillow==10.1.0",
    "beautifulsoup4==4.12.2",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "python-multipart==0.0.6",
    "aiohttp==3.9.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

